requirements.txt
httpx>=0.24.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
aiosqlite>=0.19.0
sqlalchemy[asyncio]>=2.0.0
thefuzz>=0.20.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
asyncpg>=0.28.0
web3>=6.0.0